[gazebo] Minor fixes to simulator display in gazebo
> Fixed some minor bugs in MapGen appimage
> Fixed camera position in mindstorm_arena.world
> Fixed arena_ground size

+++ script.sh
@@ -1,18 +1,26 @@
-# display line and print errors
-set -xe
-
-# variables
-conf='./utils/conf'
+# display errors
+set -e
 
 # map generation sub routine
 ./utils/MapGen-x86_64.AppImage
 
 # mv files in the code base
-mv ./conf ./utils/
-mv ./map.sdf ./utils/
-cp ./utils/map.sdf ./src/simulator/models/mindstorm_map/
+if test -f ./conf; then
+    mv ./conf ./utils/
+fi
 
-# build source and run
+if test -f ./map.sdf; then
+    mv ./map.sdf ./utils/
+    cp ./utils/map.sdf ./src/simulator/models/mindstorm_map/
+fi
+
+conf='./utils/conf'
+
 colcon build
 source ./install/setup.sh
-ros2 launch simulator sim.launch.py $(<"$conf")
+
+if test -f "${conf}"; then
+    ros2 launch simulator sim.launch.py $(<"$conf")
+else
+    echo "default"
+fi
